Initialize instrumentation
